LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focus-keyboard-js-expected.txt



FAIL focus-keyboard-js assert_equals: #text should not be selected expected 3 but got 0
